Digit_Detektor/requirements.txt
numpy
Pillow
joblib
matplotlib
opencv-python
scikit-learn
